Security

+++ admin/includes/view_all_users.php
@@ -42,27 +42,33 @@
         echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
         echo "</tr>";
     }
-    ?> 
+    ?>
     </tbody>
 </table>
 
 <?php
 // Delete User
+
 if (isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: users.php");
+
+    if (isset($_SESSION['user_role'])) {
+        if ($_SESSION['user_role'] == 'admin') {
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            $delete_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }
+    }
 }
 
-if (isset($_GET['change_to_admin'])){
+if (isset($_GET['change_to_admin'])) {
     $the_user_id = $_GET['change_to_admin'];
     $query = "UPDATE users SET user_role = 'admin' WHERE user_id = {$the_user_id}";
     $change_to_admin_query = mysqli_query($connection, $query);
     header("Location: users.php");
 }
 
-if (isset($_GET['change_to_subscriber'])){
+if (isset($_GET['change_to_subscriber'])) {
     $the_user_id = $_GET['change_to_subscriber'];
     $query = "UPDATE users SET user_role = 'subscriber' WHERE user_id = {$the_user_id}";
     $change_to_sub_query = mysqli_query($connection, $query);
